+++ app/Enums/ResponseStatusEnum.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace App\Enums;
+
+class ResponseStatusEnum
+{
+    const SUCCESS = 'SUCCESS';
+    const ERROR = 'ERROR';
+}