+++ app/Helpers/APIHelper.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Helpers;
+
+class APIHelper
+{
+    static function getHeaders() : array {
+        return [
+            'Accept' => "application/json",
+            'Content-Type' => "application/json",
+        ];
+    }
+}