+++ app/Http/Middleware/AdminOnlyMiddleware.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace App\Http\Middleware;
+
+use Closure;
+use Illuminate\Container\Attributes\Auth;
+use Illuminate\Http\Request;
+use Symfony\Component\HttpFoundation\Response;
+
+class AdminOnlyMiddleware
+{
+    /**
+     * Handle an incoming request.
+     *
+     * @param  \Closure(\Illuminate\Http\Request): (\Symfony\Component\HttpFoundation\Response)  $next
+     */
+    public function handle(Request $request, Closure $next): Response
+    {
+        if ($request->user()->is_admin == false) {
+            return response()->json([
+                'message' => 'user role is unauthorized',
+            ], Response::HTTP_FORBIDDEN);
+        }
+
+        return $next($request);
+    }
+}