+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [0.2.0] - 05-05-2025
 
 -   add feature: login, logout, get logged user
+-   add feature admin: CRUD services
 
 ## [0.1.0] - 05-05-2025